src/ALBEF/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
docker-pycreds==0.4.0
filelock==3.10.7
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.0.12
idna==3.3
importlib-metadata==6.1.0
joblib==1.1.0
jsonlines==3.0.0
numpy==1.21.6
opencv-python==4.7.0.72
packaging==20.9
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
sentry-sdk==1.5.6
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
termcolor==1.1.0
timm==0.4.9
tokenizers==0.10.3
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.62.3
transformers==4.8.1
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.10
yaspin==2.1.0
zipp==3.15.0
